+++ src/services/auth.ts
@@ -1,10 +1,6 @@
 import { eq, sql } from "drizzle-orm";
 import { db } from "../db/client";
-import { apiKeys, identities, users } from "../db/schema/auth";
-import { HTTPException } from "hono/http-exception";
-import { providers } from "@db/schema/auth";
-
-type NewUser = typeof users.$inferInsert;
+import { apiKeys } from "../db/schema/auth";
 
 /**
  * Retrieves and validates an API key from the database
@@ -37,37 +33,4 @@ export const retrieveApiKey = async (key: string) => {
     return true;
   }
   return false;
-};
-
-export const createUser = async (newUser: NewUser) => {
-  try {
-    const data = await db
-      .insert(users)
-      .values(newUser)
-      .returning({ id: users.id });
-    return data[0];
-  } catch (e) {
-    console.error(`Failed to create user: ${e}`);
-    throw new HTTPException(501, { message: "Failed to create user" });
-  }
-};
-
-export const createIdentity = async (
-  userId: string,
-  provider: (typeof providers.enumValues)[0]
-) => {
-  try {
-    const data = await db
-      .insert(identities)
-      .values({
-        id: crypto.randomUUID(),
-        userId,
-        provider,
-      })
-      .returning();
-    return data[0];
-  } catch (e) {
-    console.error(`Failed to create identity: ${e}`);
-    throw new HTTPException(501, { message: "Failed to create identity" });
-  }
 };
